añadidos cambios para mostrar quienes somos la descripcion de la pagina y zonas ademas añaddas imagnes de datos iniciales y dats iniciales. Esta parte aún hay que mehorarla ya que hay dup de registros y se muestran muchos en las zonas. Además hay que arreglar los menus para que muestren los registros correctamente con las zonas

+++ imagenes.js
@@ -82,7 +82,12 @@
     }
     
     interfaceIMG.borrarImagen = function(_img_path) {
-      fs.unlink(_img_path);
+      if(_img_path != 'uploads/originalImages/anuncia_aqui.png' && _img_path != 'uploads/anuncia_aqui.png' && _img_path != './uploads/noticias/publirreportaje.jpg') {
+        console.log("borrando " + _img_path);
+        fs.unlink(_img_path);
+      } else  {
+        console.log("no se borra " + _img_path ) 
+      }
     }
   
   return interfaceIMG;
